requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication & Security
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.1

# HTTP Client for OME API
requests==2.31.0

# XML Processing
xmltodict==0.13.0
lxml==5.0.0

# Environment Variables
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
PyYAML==6.0.1
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0
